+++ app/src/org/commcare/connect/network/personalId/PersonalIdApiService.java
@@ -44,6 +44,10 @@ public interface PersonalIdApiService {
     Call<ResponseBody> confirmBackupCode(@Header("Authorization") String token,
                                          @Body Map<String, String> confirmBackupCodeRequest);
 
+    @POST(PersonalIdApiEndpoints.SET_BACKUP_CODE)
+    Call<ResponseBody> setBackupCode(@Header("Authorization") String token,
+                                     @Body Map<String, String> body);
+
     @GET(PersonalIdApiEndpoints.CREDENTIALS)
     Call<ResponseBody> retrieveCredentials(@Header("Authorization") String token);