+++ app/unit-tests/src/org/commcare/android/util/ConnectTestUtils.kt
@@ -97,7 +97,7 @@ object ConnectTestUtils {
         createConnectDbFile()
         val user = ConnectUserRecord("", "", "", "Test User", "", Date(), null, false, "", hasConnectAccess)
         user.updateConnectToken("test-connect-token", daysFromNow(1))
-        ConnectUserDatabaseUtil.storeUser(context, user)
+        ConnectUserDatabaseUtil.storeUser(user)
         PersonalIdManager.getInstance().init(context)
     }
 
@@ -129,6 +129,5 @@ object ConnectTestUtils {
                 isLearning,
             ).apply { setLastUpdate(Date()) }
 
-    private fun <T : Persistable> connectStorage(clazz: Class<T>) =
-        ConnectDatabaseHelper.getConnectStorage(ApplicationProvider.getApplicationContext(), clazz)
+    private fun <T : Persistable> connectStorage(clazz: Class<T>) = ConnectDatabaseHelper.getConnectStorage(clazz)
 }