+++ app/unit-tests/src/org/commcare/connect/network/personalId/parser/LinkHqWorkerResponseParserTest.kt
@@ -75,7 +75,7 @@ class LinkHqWorkerResponseParserTest {
 
         // Assert
         connectAppDatabaseMock.verify(
-            { ConnectAppDatabaseUtil.storeApp(context, appRecord) },
+            { ConnectAppDatabaseUtil.storeApp(appRecord) },
             times(1),
         )
     }